Managed to have the delete button working, but ..
With the help of someone else. Anyways, I now know that whenever I dynamically create a new element in the DOM, and I want to target that specific element, then I should add the event listener in the same function in which I create the element. Doing it that way omits the check if element is in the DOM.